Automations for shared board
This is a first pass at a shared board with automations. The guiding
principle for the main board is that things shouldn't stay on the board
for "too long", whatever that means for us. Things that are not likely
to make progress in the near future, such as a PR that had changes
requested, but the author has gone silent for some period of time,
should be in "Backlog" or "Stalled".

This way, the board remains focused and easier to see all the concrete
tasks in motion, and we can have separate views for looking over
items that are not immediately actionable.